Bring ROADMAP.md up to date with current practices
This file feels misleading because it contains a lot of information from 2016 and earlier, and does not direct the reader clearly to the places like OKRs, Waffle, and REQUIREMENTS.md where the most recent action is happening.  My proposal is to replace it with this simplified file.  (Alternatively, we could delete ROADMAP.md entirely?)